+++ app/uk/gov/hmrc/tai/modules/LocalGuiceModule.scala
@@ -22,8 +22,7 @@ import play.api.{Configuration, Environment}
 import uk.gov.hmrc.auth.core.AuthorisedFunctions
 import uk.gov.hmrc.tai.auth.MicroserviceAuthorisedFunctions
 import uk.gov.hmrc.tai.config.ApplicationStartUp
-import uk.gov.hmrc.tai.connectors.cache.{CachingIabdConnector, DefaultIabdConnector, IabdConnector}
-import uk.gov.hmrc.tai.connectors.{CachingRtiConnector, DefaultRtiConnector, RtiConnector}
+import uk.gov.hmrc.tai.connectors.{CachingIabdConnector, CachingRtiConnector, DefaultIabdConnector, DefaultRtiConnector, IabdConnector, RtiConnector}
 import uk.gov.hmrc.tai.model.helpers.IncomeHelper
 import uk.gov.hmrc.tai.repositories.cache.APICacheRepository
 import uk.gov.hmrc.tai.service.{LockService, LockServiceImpl}
