+++ test/uk/gov/hmrc/tai/connectors/IabdConnectorSpec.scala
@@ -22,7 +22,6 @@ import play.api.libs.json.{JsNull, JsObject, Json}
 import play.api.test.FakeRequest
 import uk.gov.hmrc.http.{BadRequestException, HeaderNames, HttpException, NotFoundException}
 import uk.gov.hmrc.tai.config.NpsConfig
-import uk.gov.hmrc.tai.connectors.cache.{DefaultIabdConnector, IabdConnector}
 import uk.gov.hmrc.tai.controllers.predicates.AuthenticatedRequest
 import uk.gov.hmrc.tai.model.IabdUpdateAmountFormats
 import uk.gov.hmrc.tai.model.domain.formatters.IabdDetails
