+++ test/uk/gov/hmrc/tai/connectors/cache/CachingRtiConnectorSpec.scala
@@ -30,7 +30,7 @@ import uk.gov.hmrc.mongo.cache.DataKey
 import uk.gov.hmrc.mongo.lock.MongoLockRepository
 import uk.gov.hmrc.mongoFeatureToggles.services.FeatureFlagService
 import uk.gov.hmrc.tai.auth.MicroserviceAuthorisedFunctions
-import uk.gov.hmrc.tai.connectors.{CachingRtiConnector, ConnectorBaseSpec, RtiConnector}
+import uk.gov.hmrc.tai.connectors.{CachingIabdConnector, CachingRtiConnector, ConnectorBaseSpec, DefaultIabdConnector, IabdConnector, RtiConnector}
 import uk.gov.hmrc.tai.model.domain.{AnnualAccount, Available, FourWeekly, Payment, RtiPaymentsForYearError, ServiceUnavailableError}
 import uk.gov.hmrc.tai.model.tai.TaxYear
 import uk.gov.hmrc.tai.repositories.cache.{APICacheRepository, TaiSessionCacheRepository}
