+++ test/uk/gov/hmrc/tai/repositories/IabdRepositorySpec.scala
@@ -21,7 +21,8 @@ import play.api.libs.json.Json
 import play.api.test.FakeRequest
 import uk.gov.hmrc.http.NotFoundException
 import uk.gov.hmrc.tai.config.CacheMetricsConfig
-import uk.gov.hmrc.tai.connectors.cache.{Caching, IabdConnector}
+import uk.gov.hmrc.tai.connectors.IabdConnector
+import uk.gov.hmrc.tai.connectors.cache.Caching
 import uk.gov.hmrc.tai.controllers.predicates.AuthenticatedRequest
 import uk.gov.hmrc.tai.metrics.Metrics
 import uk.gov.hmrc.tai.model.domain.formatters.IabdDetails
